requirements_api.txt
black
pylint
pytest
pytest-cov
scikit-learn
tensorflow==2.20.0
fastapi
uvicorn[standard]
pillow
python-multipart
jinja2
pydantic
pyyaml